fix: autoscroll, header, reminders
