+++ android/app/src/main/java/com/WahegurooNetwork/SundarGutka/MainActivity.java
@@ -17,7 +17,7 @@ public class MainActivity extends ReactActivity {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
-    SplashScreen.show(this);
+    SplashScreen.show(this,true);
     super.onCreate(savedInstanceState);
   }
 }
