+++ app.js
@@ -2,9 +2,10 @@ import React, { useEffect } from "react";
 import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
 import ErrorBoundary from "react-native-error-boundary";
+import notifee, { EventType } from "@notifee/react-native";
 import crashlytics from "@react-native-firebase/crashlytics";
 import SplashScreen from "react-native-splash-screen";
-import { createStore, errorHandler, FallBack } from "@common";
+import { createStore, errorHandler, FallBack, resetBadgeCount, navigateTo } from "@common";
 import Navigation from "./src/navigation";
 import { allowTracking } from "./src/common/analytics";
 
@@ -21,6 +22,15 @@ const App = () => {
   useEffect(() => {
     enableCrashlytics();
     allowTracking();
+  }, []);
+
+  useEffect(() => {
+    return notifee.onForegroundEvent(({ type, detail }) => {
+      resetBadgeCount();
+      if (type === EventType.PRESS) {
+        navigateTo(detail);
+      }
+    });
   }, []);
 
   return (
