+++ index.js
@@ -5,16 +5,14 @@
 import { AppRegistry } from "react-native";
 import notifee, { EventType } from "@notifee/react-native";
 import App from "./app";
+import { resetBadgeCount } from "./src/common/notifications";
+import { navigateTo } from "./src/common/rootNavigation";
 import { name as appName } from "./app.json";
-import { navigateTo } from "./src/navigation/util";
 
 notifee.onBackgroundEvent(async ({ type, detail }) => {
-  const { notification } = detail;
-
-  // Check if the user pressed the "Mark as read" action
-  if (type === EventType.ACTION_PRESS) {
+  resetBadgeCount();
+  if (type === EventType.PRESS) {
     navigateTo(detail);
-    await notifee.cancelNotification(notification.id);
   }
 });
