+++ src/ReaderScreen/hooks/index.js
@@ -1,6 +1,5 @@
 import useBookmarks from "./useBookmarks";
 import useFetchShabad from "./useFetchShabad";
 import usePrevious from "./usePrevious";
-import useAnimationHeadFoot from "./useAnimationHeadFoot";
 
-export { useBookmarks, useFetchShabad, usePrevious, useAnimationHeadFoot };
+export { useBookmarks, useFetchShabad, usePrevious };
