+++ src/navigation/index.jsx
@@ -1,10 +1,9 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { useSelector } from "react-redux";
 import { Icon } from "@rneui/themed";
-import notifee, { EventType } from "@notifee/react-native";
-import { constant, colors, navigationRef, navigate, resetBadgeCount } from "@common";
+import { colors, navigationRef } from "@common";
 import HomeScreen from "../HomeScreen";
 import Reader from "../ReaderScreen";
 import Settings from "../Settings";
@@ -29,53 +28,6 @@ const Navigation = () => {
   const isNightMode = useSelector((state) => state.isNightMode);
   const settingsStyle = SettingsStyle(isNightMode);
   const { headerTitleStyle, headerStyle } = settingsStyle;
-
-  const navigateTo = (incoming) => {
-    const { data } = incoming.notification;
-    const params = { key: `Reader-${data.id}`, params: { id: data.id, title: data.gurmukhi } };
-    navigate(constant.READER, params);
-  };
-
-  useEffect(() => {
-    async function setupNotifications() {
-      // Correctly handle the initial notification promise
-      const initialNotification = await notifee.getInitialNotification();
-      if (initialNotification) {
-        resetBadgeCount();
-      }
-
-      // Setting up event listeners
-      const unsubscribeForeground = notifee.onForegroundEvent(({ type, detail }) => {
-        switch (type) {
-          case EventType.PRESS:
-            navigateTo(detail);
-            resetBadgeCount();
-            break;
-          default:
-            resetBadgeCount();
-        }
-      });
-
-      const unsubscribeBackground = notifee.onBackgroundEvent(({ type, detail }) => {
-        switch (type) {
-          case EventType.PRESS:
-            navigateTo(detail);
-            resetBadgeCount();
-            break;
-          default:
-            resetBadgeCount();
-        }
-      });
-
-      // Cleanup function to unsubscribe events
-      return () => {
-        unsubscribeForeground();
-        unsubscribeBackground();
-      };
-    }
-
-    setupNotifications();
-  }, []);
 
   return (
     <NavigationContainer ref={navigationRef}>
